% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbrdata-package.R
\docType{package}
\name{msigdbrdata-package}
\alias{msigdbrdata-package}
\title{msigdbrdata: MSigDB Data Frame}
\description{
Provides the Molecular Signatures Database (MSigDB) gene sets (Subramanian et al. 2005 \doi{10.1073/pnas.0506580102}, Liberzon et al. 2015 \doi{10.1016/j.cels.2015.12.004}, Castanza et al. 2023 \doi{10.1038/s41592-023-02014-7}) in an R data frame. Serves as the data source for 'msigdbr' \doi{10.32614/CRAN.package.msigdbr} (split into a separate package due to CRAN size restrictions).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/igordot/msigdbrdata}
  \item Report bugs at \url{https://github.com/igordot/msigdbrdata/issues}
}

}
\author{
\strong{Maintainer}: Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})

}
\keyword{internal}
